8.29
子弹类问题

+++ Classes/bullet/Bullet.h
@@ -2,7 +2,7 @@
 #define _BULLET_H_
 
 #include"cocos2d.h"
-#include"Enemy.h"
+
 USING_NS_CC;
 
 class Bullet :public Sprite {
@@ -12,13 +12,13 @@ public:
 	int Shoot(int tagoftower, int Tag);
 
 	//bool IfKill();
-	int setDamage(int D);
+	int setDamage(double D);
 protected:
 	bool kill = 0;
 	Scene * rScene;
 	void Hit(float dt);
 	void clear(float dt);
-	int damage = 25;
+	double damage = 25;
 	int tagOfEnemy;
 	int tagOfTower;
 };
